Change to float

+++ cla_backend/apps/call_centre/serializers.py
@@ -35,7 +35,7 @@ from legalaid.models import PersonalDetails
 class PropertySerializer(PropertySerializerBase):
     value = serializers.IntegerField(required=True)
     mortgage_left = serializers.IntegerField(required=True)
-    share = serializers.IntegerField(required=True)
+    share = serializers.FloatField(required=True)
 
     class Meta(PropertySerializerBase.Meta):
         fields = ("value", "mortgage_left", "share", "id", "disputed", "main")
